requirements.txt
pyautogui
pyperclip
pillow
openai
google-generativeai
opencv-python
psutil
pywin32
screeninfo
